@@ -32,7 +32,7 @@
     "@vitejs/plugin-react": "^4.4.1",
     "cross-env": "^7.0.3",
     "csstype": "^3.1.3",
-    "typescript": "^5.8.3",
+    "typescript": "^5.9.2",
     "vite": "^6.3.2"
   }
 }
