@@ -20,7 +20,7 @@
     "@types/react-dom": "^18.2.19",
     "@vitejs/plugin-react": "^4.2.1",
     "cross-env": "^7.0.3",
-    "typescript": "^5.2.2",
+    "typescript": "^5.9.2",
     "vite": "^5.2.11"
   }
 }
