@@ -30,7 +30,7 @@
     "@vitejs/plugin-react": "^4.2.1",
     "concurrently": "^8.2.2",
     "cross-env": "^7.0.3",
-    "typescript": "^5.2.2",
+    "typescript": "^5.9.2",
     "vite": "^5.1.4"
   }
 }
