@@ -18,7 +18,7 @@
   },
   "devDependencies": {
     "cross-env": "^7.0.3",
-    "typescript": "^5.2.2",
+    "typescript": "^5.9.2",
     "vite": "^5.2.11"
   }
 }
