@@ -138,8 +138,8 @@
     "@types/react": "^18.0.8",
     "@types/react-dom": "^18.0.3",
     "@types/trusted-types": "^2.0.7",
-    "@typescript-eslint/eslint-plugin": "^7.8.0",
-    "@typescript-eslint/parser": "^7.8.0",
+    "@typescript-eslint/eslint-plugin": "^8.39.1",
+    "@typescript-eslint/parser": "^8.39.1",
     "child-process-promise": "^2.2.1",
     "chokidar": "^3.5.2",
     "concurrently": "^8.2.2",
@@ -177,7 +177,7 @@
     "lint-staged": "^11.1.0",
     "minimist": "^1.2.5",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.8.8",
+    "prettier": "^3.6.2",
     "prettier-plugin-hermes-parser": "^0.26.0",
     "prettier-plugin-organize-attributes": "^0.0.5",
     "prettier-plugin-tailwindcss": "^0.4.1",
@@ -188,8 +188,8 @@
     "ts-jest": "^29.1.2",
     "ts-morph": "^25.0.1",
     "ts-node": "^10.9.1",
-    "typedoc": "^0.25.12",
-    "typescript": "^5.4.5",
+    "typedoc": "^0.28.10",
+    "typescript": "^5.9.2",
     "vite": "^5.2.11",
     "vite-tsconfig-paths": "^5.1.4",
     "vitest": "^3.2.4"
