@@ -90,7 +90,12 @@ module.exports = {
         '@typescript-eslint/no-this-alias': OFF,
         '@typescript-eslint/no-unused-vars': [
           ERROR,
-          {args: 'none', argsIgnorePattern: '^_', varsIgnorePattern: '^_'},
+          {
+            args: 'none',
+            argsIgnorePattern: '^_',
+            caughtErrorsIgnorePattern: '^_',
+            varsIgnorePattern: '^_',
+          },
         ],
         'header/header': [2, 'scripts/www/headerTemplate.js'],
       },
