@@ -149,12 +149,12 @@ describe('LexicalHeadlessEditor', () => {
       editor.dispatchCommand(CONTROLLED_TEXT_INSERTION_COMMAND, 'foo');
     });
 
-    expect(onUpdate).toBeCalled();
-    expect(onCommand).toBeCalledWith('foo', expect.anything());
-    expect(onTransform).toBeCalledWith(
+    expect(onUpdate).toHaveBeenCalled();
+    expect(onCommand).toHaveBeenCalledWith('foo', expect.anything());
+    expect(onTransform).toHaveBeenCalledWith(
       expect.objectContaining({__type: 'paragraph'}),
     );
-    expect(onTextContent).toBeCalledWith('Helloworld');
+    expect(onTextContent).toHaveBeenCalledWith('Helloworld');
   });
 
   it('can preserve selection for pending editor state (within update loop)', async () => {
