@@ -69,7 +69,7 @@ function tryToPositionRange(
   try {
     range.setStart(anchorNode, startOffset);
     range.setEnd(anchorNode, endOffset);
-  } catch (error) {
+  } catch (_error) {
     return false;
   }
 
