@@ -66,7 +66,7 @@ export function useYjsCollaboration(
   const disconnect = useCallback(() => {
     try {
       provider.disconnect();
-    } catch (e) {
+    } catch (_e) {
       // Do nothing
     }
   }, [provider]);
