@@ -99,7 +99,7 @@ export function createDOMRange(
   try {
     range.setStart(anchorDOM, anchorOffset);
     range.setEnd(focusDOM, focusOffset);
-  } catch (e) {
+  } catch (_e) {
     return null;
   }
 
