@@ -11,6 +11,7 @@ import {
   $cloneWithProperties,
   $createParagraphNode,
   $getAdjacentChildCaret,
+  $getAdjacentSiblingOrParentSiblingCaret,
   $getCaretInDirection,
   $getChildCaret,
   $getChildCaretOrSelf,
@@ -41,7 +42,6 @@ import {
   type NodeCaret,
   type NodeKey,
   PointCaret,
-  RootMode,
   type SiblingCaret,
   SplitAtPointCaretNextOptions,
   StateConfig,
@@ -69,6 +69,7 @@ export {default as mergeRegister} from './mergeRegister';
 export {default as positionNodeOnRange} from './positionNodeOnRange';
 export {default as selectionAlwaysOnDisplay} from './selectionAlwaysOnDisplay';
 export {
+  $getAdjacentSiblingOrParentSiblingCaret,
   $splitNode,
   isBlockDomNode,
   isHTMLAnchorElement,
@@ -870,35 +871,6 @@ export function $unwrapNode(node: ElementNode): void {
     1,
     node.getChildren(),
   );
-}
-
-/**
- * Returns the Node sibling when this exists, otherwise the closest parent sibling. For example
- * R -> P -> T1, T2
- *   -> P2
- * returns T2 for node T1, P2 for node T2, and null for node P2.
- * @param node LexicalNode.
- * @returns An array (tuple) containing the found Lexical node and the depth difference, or null, if this node doesn't exist.
- */
-export function $getAdjacentSiblingOrParentSiblingCaret<
-  D extends CaretDirection,
->(
-  startCaret: NodeCaret<D>,
-  rootMode: RootMode = 'root',
-): null | [NodeCaret<D>, number] {
-  let depthDiff = 0;
-  let caret = startCaret;
-  let nextCaret = $getAdjacentChildCaret(caret);
-  while (nextCaret === null) {
-    depthDiff--;
-    nextCaret = caret.getParentCaret(rootMode);
-    if (!nextCaret) {
-      return null;
-    }
-    caret = nextCaret;
-    nextCaret = $getAdjacentChildCaret(caret);
-  }
-  return nextCaret && [nextCaret, depthDiff];
 }
 
 /**
