@@ -277,9 +277,8 @@ function updateCursor(
     const style = `position:absolute;top:${top}px;left:${left}px;height:${selectionRect.height}px;width:${selectionRect.width}px;pointer-events:none;z-index:5;`;
     selection.style.cssText = style;
 
-    (
-      selection.firstChild as HTMLSpanElement
-    ).style.cssText = `${style}left:0;top:0;background-color:${color};opacity:0.3;`;
+    (selection.firstChild as HTMLSpanElement).style.cssText =
+      `${style}left:0;top:0;background-color:${color};opacity:0.3;`;
 
     if (i === selectionRectsLength - 1) {
       if (caret.parentNode !== selection) {
