@@ -308,8 +308,8 @@ function isLastChildLineBreakOrDecorator(
         return $isLineBreakNode(node)
           ? 'line-break'
           : $isDecoratorNode(node) && node.isInline()
-          ? 'decorator'
-          : null;
+            ? 'decorator'
+            : null;
       }
     }
     return 'empty';
