@@ -178,7 +178,7 @@ export function isSelectionWithinEditor(
       !isSelectionCapturedInDecoratorInput(anchorDOM) &&
       getNearestEditorFromDOMNode(anchorDOM) === editor
     );
-  } catch (error) {
+  } catch (_error) {
     return false;
   }
 }
@@ -1340,8 +1340,8 @@ export function getDOMOwnerDocument(
   return isDOMDocumentNode(target)
     ? target
     : isHTMLElement(target)
-    ? target.ownerDocument
-    : null;
+      ? target.ownerDocument
+      : null;
 }
 
 export function scrollIntoViewIfNeeded(
