@@ -1206,8 +1206,8 @@ describe('LexicalCaret', () => {
                     anchorBias === 'outside' && focusBias === 'outside'
                       ? []
                       : (anchorBias === 'inside') === (direction === 'next')
-                      ? [{caret: {offset: 0}, distance: size}]
-                      : [{caret: {offset: size}, distance: -size}],
+                        ? [{caret: {offset: 0}, distance: size}]
+                        : [{caret: {offset: size}, distance: -size}],
                   );
                   const resultRange = $removeTextFromCaretRange(range);
                   $setSelection(null);
